% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{availablePalette}
\alias{availablePalette}
\title{Available color palette}
\usage{
availablePalette()
}
\value{
a character vector contain available color palettes.
}
\description{
List all the available color palettes.
}
\examples{
availablePalette()
}
