% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colorNames}
\alias{colorNames}
\title{available color variable}
\usage{
colorNames()
}
\value{
a character vector contain safe colors.
}
\description{
export available color names
}
\examples{
colorNames()
}
