% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{paletteMartin}
\alias{paletteMartin}
\alias{Green2Magenta16Steps}
\alias{Blue2DarkRed12Steps}
\alias{Blue2DarkRed18Steps}
\alias{Blue2OrangeRed14Steps}
\alias{Blue2DarkOrange12Steps}
\alias{Blue2DarkOrange18Steps}
\alias{Blue2Green14Steps}
\alias{Brown2Blue10Steps}
\alias{Brown2Blue12Steps}
\alias{Blue2Gray8Steps}
\alias{Blue2Orange8Steps}
\alias{Blue2Orange10Steps}
\alias{Blue2Orange12Steps}
\alias{ModifiedSpectralScheme11Steps}
\alias{LightBlue2DarkBlue7Steps}
\alias{LightBlue2DarkBlue10Steps}
\alias{PairedColor12Steps}
\alias{SteppedSequential5Steps}
\title{Palette for color blindness}
\format{An object of class \code{character} of length 15.}
\source{
http://mkweb.bcgsc.ca/biovis2012/

http://geog.uoregon.edu/datagraphics/color_scales.htm
}
\usage{
paletteMartin

Green2Magenta16Steps

Blue2DarkRed12Steps

Blue2DarkRed18Steps

Blue2OrangeRed14Steps

Blue2DarkOrange12Steps

Blue2DarkOrange18Steps

Blue2Green14Steps

Brown2Blue10Steps

Brown2Blue12Steps

Blue2Gray8Steps

Blue2Orange8Steps

Blue2Orange10Steps

Blue2Orange12Steps

ModifiedSpectralScheme11Steps

LightBlue2DarkBlue7Steps

LightBlue2DarkBlue10Steps

PairedColor12Steps

SteppedSequential5Steps
}
\description{
The palette could be used for heatmap or pie graph
}
\details{
The names of the palette is approximal color name.

Green2Magenta16Steps: Useful for generic diverging data.

Blue2DarkRed12/18Steps: Useful for temperature-like data, 
with a subjective interpretation (blue=cold, red=hot)
Blue2OrangeRed14Steps: Useful as an alternative to the red/blue temperature scale.

Blue2DarkOrange12/18Steps: Useful for data without a specific subjective color association.

Blue2Green14Steps: Useful for data with a winter (blue) vs. summer (green) association.

Brown2Blue10/12Steps: Useful for data with a dry (brown) vs. wet (blue) association.

Blue2Gray8Steps: Useful in particular for diverging data like cloudiness anomalies.

Blue2Orange8/10/12Steps: Useful for data like sea-level pressure, 
with ansubjective association (blue=low, wet, orange=high, dry)

ModifiedSpectralScheme11Steps:	An alternative to the spectral scheme (no green)

LightBlue2DarkBlue7/10Steps: Useful for precipitation-like data.

PairedColor12Steps: Attempt at a categorical color scale with colors
that may be distinguishable to all viewers

SteppedSequential5Steps: Useful for portraying levels-within-categories
}
\examples{
paletteMartin
Green2Magenta16Steps
Blue2DarkRed12Steps
Blue2DarkRed18Steps
Blue2OrangeRed14Steps
Blue2DarkOrange12Steps
Blue2DarkOrange18Steps
Blue2Green14Steps
Brown2Blue10Steps
Brown2Blue12Steps
Blue2Gray8Steps
Blue2Orange8Steps
Blue2Orange10Steps
Blue2Orange12Steps
ModifiedSpectralScheme11Steps
LightBlue2DarkBlue7Steps
LightBlue2DarkBlue10Steps
PairedColor12Steps
SteppedSequential5Steps
}
\references{
Light A, Bartlein PJ (2004). The End of the Rainbow? Color Schemes for Improved Data Graphics.
EOS Transactions of the American Geophysical Union, 85(40), 385.
}
\keyword{datasets}
