% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdPlot.R
\name{cvdPlot}
\alias{cvdPlot}
\title{Show color-deficiency simulations of a plot}
\usage{
cvdPlot(plot = last_plot())
}
\arguments{
\item{plot}{The grob to be plotted.}
}
\description{
Plot the color-deficiency simulations for ggplot grob.
}
\details{
This function is modified from https://github.com/clauswilke/colorblindr
}
\examples{
colorBlindness:::cvdPlot(displayColors(safeColors))
colorBlindness:::cvdPlot(displayColors(paletteMartin))
}
