% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdSimulator.R
\name{cvdSimulator}
\alias{cvdSimulator}
\title{simulate color vision deficiency}
\usage{
cvdSimulator(col, type = "deuteranope")
}
\arguments{
\item{col}{character. A vector of colors.}

\item{type}{Deficiency type, "protanope" or "deuteranope"}
}
\value{
colors.
}
\description{
Transformation of R colors by simulating color
 vision deficiencies.
}
\details{
Here use Vienot's methods but not Gustavo's methods 
(implemented in colorspace::simulate_cvd).
}
\examples{
cvdSimulator(safeColors)
}
\references{
Vienot, F., Brettel, H. and Mollon, J.D. (1999) 
<doi:10.1002/(SICI)1520-6378(199908)24:4%3C243::AID-COL5%3E3.0.CO;2-3>
Sharma, G., Wu, W. and Dalal, E.N. (2005) <doi:10.1002/col.20070>
}
