% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{displayAvailablePalette}
\alias{displayAvailablePalette}
\title{Display available palette}
\usage{
displayAvailablePalette(...)
}
\arguments{
\item{...}{parameters could be used by \link[ggplot2:geom_raster]{geom_tile}.}
}
\value{
an \link[ggplot2:ggplot]{ggplot} object
}
\description{
Display all the available color palettes.
}
\examples{
displayAvailablePalette()
}
