% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{palette15}
\alias{palette15}
\title{15-color palette for color blindness}
\format{An object of class \code{character} of length 15.}
\usage{
palette15
}
\description{
The 15 color palette could be used for heatmap or pie graph
}
\details{
The names of the palette is approximal color name.
}
\references{
http://mkweb.bcgsc.ca/biovis2012/
}
\keyword{datasets}
