% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacePlotColor.R
\name{replacePlotColor}
\alias{replacePlotColor}
\title{replace the colors for plots}
\usage{
replacePlotColor(plot)
}
\arguments{
\item{plot}{The grob to be plotted.}
}
\value{
an object of gtable.
}
\description{
replace the colors of plots to meet 
the requirment of publication.
Replacing red with magenta or green with turquoise.
Replacing all the colored symbols in the legend.
}
\examples{
replacePlotColor(displayColors(c("Red", "Green", "blue")))
}
