% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{BLACK}
\alias{BLACK}
\alias{ORANGE}
\alias{SKY_BLUE}
\alias{BLUISH_GREEN}
\alias{YELLOW}
\alias{BLUE}
\alias{VERMILLION}
\alias{REDDISH_PURPLE}
\alias{safeColors}
\title{safe colors}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 8.
}
\usage{
BLACK

ORANGE

SKY_BLUE

BLUISH_GREEN

YELLOW

BLUE

VERMILLION

REDDISH_PURPLE

safeColors
}
\description{
color blindness safe colors
}
\examples{
safeColors
}
\references{
Wong, B. (2011) <doi:10.1038/nmeth.1642> 
Wong, B. (2011) <doi:10.1038/nmeth.1618>
}
\keyword{datasets}
