% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdPlot.R
\name{cvdPlot}
\alias{cvdPlot}
\title{Show color-deficiency simulations of a plot}
\usage{
cvdPlot(
  plot = last_plot(),
  layout = c("origin", "deuteranope", "protanope", "desaturate")
)
}
\arguments{
\item{plot}{The grob to be plotted.}

\item{layout}{The sub-figure types. the choices are origin, deuteranope,
protanope, desaturate, and enhanced, enhanced.deuteranope, 
enhanced.protanope, enhanced.desaturate.}
}
\value{
An object of ggplot.
}
\description{
Plot the color-deficiency simulations for ggplot grob.
}
\details{
This function is modified from <https://github.com/clauswilke/colorblindr>
}
\examples{
cvdPlot(displayColors(safeColors))
cvdPlot(displayColors(paletteMartin))
}
