% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdSimulator.R
\name{cvdSimulator}
\alias{cvdSimulator}
\title{simulate color vision deficiency}
\usage{
cvdSimulator(col, type = c("deuteranope", "protanope", "desaturate"))
}
\arguments{
\item{col}{character. A vector of colors.}

\item{type}{Deficiency type, "protanope" or "deuteranope"}
}
\description{
Transformation of R colors by simulating color vision deficiencies.
}
\details{
Here use Vienot's methods but not Gustavo's methods ( implemented in colorspace::simulate_cvd).
}
\examples{
cvdSimulator(safeColors)
}
\references{
Vienot, F., Brettel, H. and Mollon, J.D., 1999. Digital video colourmaps for 
checking the legibility of displays by dichromats. 
Color Research & Application. 24(4), pp.243-252.
}
