% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdSimulator.R
\name{cvdSimulator}
\alias{cvdSimulator}
\title{simulate color vision deficiency}
\usage{
cvdSimulator(col, type = "deuteranope")
}
\arguments{
\item{col}{character. A vector of colors.}

\item{type}{Deficiency type, "protanope" or "deuteranope"}
}
\description{
Transformation of R colors by simulating color vision deficiencies.
}
\details{
Here use Vienot's methods but not Gustavo's methods ( implemented in colorspace::simulate_cvd).
}
\examples{
cvdSimulator(safeColors)
}
\references{
Vienot, F., Brettel, H. and Mollon, J.D., 1999. Digital video colourmaps for 
checking the legibility of displays by dichromats. 
Color Research & Application. 24(4), pp.243-252.
Sharma, G., Wu, W. and Dalal, E.N., 2005. The CIEDE2000 color‐difference formula:
 Implementation notes, supplementary test data, and mathematical observations. 
 Color Research & Application 30.1, pp.21-30.
}
