% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayColors.R
\name{displayColors}
\alias{displayColors}
\title{display colors}
\usage{
displayColors(col)
}
\arguments{
\item{col}{color set to display}
}
\description{
Display the given colors
}
\examples{
displayColors(safeColors)
displayColors(palette15)
}
