% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPDFopt.R
\name{setPDFopt}
\alias{setPDFopt}
\title{Auxiliary funciton to set width of pdf for journals}
\usage{
setPDFopt(preSet = c("1col", "1.5col", "2col"))
}
\arguments{
\item{preSet}{The pre-setting of width,height,family,font for pdf.
Available choices: 1col, 1.5col, 2col.}
}
\description{
Set the pdf width and height for journals.
}
\details{
The family will be Helvetica.
The font will be 8.
The width and height will be same. 
1col=3.54 inches (9 cm.);
1.5col=5 inches (12.7 cm.); 
2col=7.25 inches (18.4 cm.).
}
\examples{
setPDFopt("1col")
}
\references{
http://www.sciencemag.org/sites/default/files/Figure_prep_guide.pdf
}
